Update Configuration class with new release
Update Configuration Class. Now the Parameter List is composed by RequestParameter. Result doesn't change, just a more correct way to code. The authenticationTokenHeader was no more a good example to use as a Header Parameter because with the new release of Swagger Authentication header can be easily implemented when coding a method and more informations about the authnetication can be specified if needed.